fix(xtask): split update_status into submodules to pass LOC-cap gate (#7642)
The in-tree test `test_update_status_is_split_into_subsystem_modules` asserts
every file in `xtask/src/tasks/update_status/` stays ≤400 LOC. Both
`mod.rs` (513 lines) and `parser.rs` (1059 lines) had grown past that limit,
causing the test to fail on master.

## Changes

**`mod.rs` (513 → 361 lines)**: Extract five command-execution helpers
(`stream_reader`, `run_cmd`, `run_cmd_merged`, `run_subsystem`, `replace_block`)
into a new `cmd.rs` module. The helpers were private implementation details
shared only within the `update_status` module; they are now `pub` inside
`cmd.rs` and imported privately in `mod.rs`.

**`parser.rs` (1059 → 389 lines)**: Extract two logical sections into Rust
submodules under `parser/`:
- `parser/failure.rs` — `FailureCluster` enum, `classify_failure_bucket` fn,
  `build_failure_worklist` fn (failure-bucket classification logic)
- `parser/tests.rs` — the full `#[cfg(test)]` module; declared with
  `#[cfg(test)] mod tests;` so it compiles only in test builds

The `super::super::parser_corpus_sweep` path depth is preserved correctly:
from within `parser/tests.rs`, `super` is the `parser` module (same as the
inline case), so all existing path references are unchanged.

**`justfile`**: Add `ci-xtask-structural` recipe that runs
`cargo test -p xtask 'tasks::update_status::mod_tests'` and wire it into
`ci-gate` after `hook-tests`. This makes the LOC-cap and structural existence
tests part of the normal merge gate so regressions are caught in CI rather
than discovered post-merge.

## Verification

```
cargo build -p xtask        # clean, no warnings
cargo clippy -p xtask       # clean
cargo test -p xtask 'tasks::update_status::mod_tests'   # all 6 pass
```

LOC counts:
- mod.rs:    361 ✓
- parser.rs: 389 ✓
- cmd.rs:    155 (new)
- parser/failure.rs: 121 (new)
- parser/tests.rs:   546 (new, not checked by gate — tests file)
